# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Callora API",
    "version": "1.0.0",
    "description": "API contract covering Callora backend billing, usage, and developer routes."
  },
  "paths": {
...
    "/api/apis": {
      "get": {
        "summary": "List public APIs",
        "description": "Returns a paginated list of public active APIs, optionally filtered by category or search term. Successful responses include a strong ETag; clients may send If-None-Match for conditional GET (304 when unchanged).",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Strong ETag from a previous listings response. When it matches the current representation, the server returns 304 Not Modified.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of public APIs",
            "headers": {
              "ETag": {
                "description": "Strong SHA-256 ETag of the JSON response body",
                "schema": {
                  "type": "string",
                  "example": "\"a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456\""
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApisListResponse"
                },
                "examples": {
                  "activeListings": {
                    "summary": "Active API listings page",
                    "value": {
                      "data": [
                        {
                          "id": 101,
                          "name": "GrantFox Scoring API",
                          "description": "Scores FWC26 grant applications against program rules.",
                          "base_url": "https://api.grantfox.example",
                          "logo_url": "https://cdn.grantfox.example/logo.png",
                          "category": "grants",
                          "status": "active",
                          "developer": {
                            "id": 11,
                            "name": "GrantFox Labs"
                          },
                          "endpoints": [
                            {
                              "id": 501,
                              "api_id": 101,
                              "path": "/applications/{applicationId}/score",
                              "method": "POST",
                              "price_per_call_usdc": "0.2500000",
                              "description": "Score a single grant application.",
                              "created_at": "2026-07-27T09:00:00.000Z",
                              "updated_at": "2026-07-27T09:00:00.000Z"
                            }
                          ]
                        },
                        {
                          "id": 102,
                          "name": "GrantFox Eligibility API",
                          "description": "Validates applicant eligibility before scoring.",
                          "base_url": "https://api.grantfox.example",
                          "logo_url": null,
                          "category": "grants",
                          "status": "active",
                          "developer": {
                            "id": 11,
                            "name": "GrantFox Labs"
                          },
                          "endpoints": [
                            {
                              "id": 502,
                              "api_id": 102,
                              "path": "/applications/{applicationId}/eligibility",
                              "method": "GET",
                              "price_per_call_usdc": "0.0500000",
                              "description": "Return eligibility checks for an application.",
                              "created_at": "2026-07-27T09:05:00.000Z",
                              "updated_at": "2026-07-27T09:05:00.000Z"
                            }
                          ]
                        }
                      ],
                      "meta": {
                        "limit": 20,
                        "hasMore": false
                      }
                    }
                  }
                }
              }
            }
          },
          "304": {
            "description": "Not Modified — the listings representation matches If-None-Match; response body is empty",
            "headers": {
              "ETag": {
                "description": "Strong ETag of the unchanged representation",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "internalServerError": {
                    "summary": "Unexpected repository error",
                    "value": {
                      "code": "INTERNAL_SERVER_ERROR",
                      "message": "Internal server error",
                      "requestId": "req-apis-list-500"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Publish a new API",
        "description": "Registers a new API with associated endpoints for the authenticated developer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiCreateRequest"
              },
              "examples": {
                "publishApi": {
                  "summary": "Publish a grant review API",
                  "value": {
                    "name": "GrantFox Review API",
                    "description": "Automates FWC26 application review workflows.",
                    "base_url": "https://api.grantfox.example",
                    "category": "grants",
                    "endpoints": [
                      {
                        "path": "/applications/{applicationId}/review",
                        "method": "POST",
                        "price_per_call_usdc": "0.3000000",
                        "description": "Run the full review pipeline for one application."
                      },
                      {
                        "path": "/applications/{applicationId}/status",
                        "method": "GET",
                        "price_per_call_usdc": "0.0200000",
                        "description": "Fetch the latest review status."
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCreateResponse"
                },
                "examples": {
                  "createdApi": {
                    "summary": "API published successfully",
                    "value": {
                      "id": 301,
                      "developer_id": 11,
                      "name": "GrantFox Review API",
                      "description": "Automates FWC26 application review workflows.",
                      "base_url": "https://api.grantfox.example",
                      "logo_url": null,
                      "category": "grants",
                      "status": "active",
                      "created_at": "2026-07-27T10:00:00.000Z",
                      "updated_at": "2026-07-27T10:00:00.000Z",
                      "endpoints": [
                        {
                          "id": 801,
                          "api_id": 301,
                          "path": "/applications/{applicationId}/review",
                          "method": "POST",
                          "price_per_call_usdc": "0.3000000",
                          "description": "Run the full review pipeline for one application.",
                          "created_at": "2026-07-27T10:00:00.000Z",
                          "updated_at": "2026-07-27T10:00:00.000Z"
                        },
                        {
                          "id": 802,
                          "api_id": 301,
                          "path": "/applications/{applicationId}/status",
                          "method": "GET",
                          "price_per_call_usdc": "0.0200000",
                          "description": "Fetch the latest review status.",
                          "created_at": "2026-07-27T10:00:00.000Z",
                          "updated_at": "2026-07-27T10:00:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload or developer profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidPayload": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "Path must start with /",
                      "requestId": "req-apis-create-400",
                      "details": [
                        {
                          "field": "endpoints.0.path",
                          "message": "Path must start with /",
                          "code": "custom"
                        }
                      ]
                    }
                  },
                  "developerProfileMissing": {
                    "summary": "Developer profile missing",
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "Developer profile not found. Create a developer profile first.",
                      "requestId": "req-apis-create-dev-missing"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Missing authentication",
                    "value": {
                      "code": "UNAUTHORIZED",
                      "message": "Unauthorized",
                      "requestId": "req-apis-create-401"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "internalServerError": {
                    "summary": "Unexpected persistence error",
                    "value": {
                      "code": "INTERNAL_SERVER_ERROR",
                      "message": "Internal server error",
                      "requestId": "req-apis-create-500"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/apis/{id}": {
      "get": {
        "summary": "Get API details",
        "description": "Returns the detailed schema and endpoints of an API by its numerical ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDetailsResponse"
                },
                "examples": {
                  "apiDetails": {
                    "summary": "Detailed API record",
                    "value": {
                      "id": 101,
                      "name": "GrantFox Scoring API",
                      "description": "Scores FWC26 grant applications against program rules.",
                      "base_url": "https://api.grantfox.example",
                      "logo_url": "https://cdn.grantfox.example/logo.png",
                      "category": "grants",
                      "status": "active",
                      "developer": {
                        "id": 11,
                        "name": "GrantFox Labs"
                      },
                      "endpoints": [
                        {
                          "id": 501,
                          "api_id": 101,
                          "path": "/applications/{applicationId}/score",
                          "method": "POST",
                          "price_per_call_usdc": "0.2500000",
                          "description": "Score a single grant application.",
                          "created_at": "2026-07-27T09:00:00.000Z",
                          "updated_at": "2026-07-27T09:00:00.000Z"
                        },
                        {
                          "id": 503,
                          "api_id": 101,
                          "path": "/applications/{applicationId}/score/explain",
                          "method": "GET",
                          "price_per_call_usdc": "0.0400000",
                          "description": "Fetch scoring rationale for the latest run.",
                          "created_at": "2026-07-27T09:10:00.000Z",
                          "updated_at": "2026-07-27T09:10:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidId": {
                    "summary": "Invalid API id",
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "id must be a positive integer",
                      "requestId": "req-api-details-400"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "API not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "apiNotFound": {
                    "summary": "API not found",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "API not found or not active",
                      "requestId": "req-api-details-404"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "internalServerError": {
                    "summary": "Unexpected lookup error",
                    "value": {
                      "code": "INTERNAL_SERVER_ERROR",
                      "message": "Internal server error",
                      "requestId": "req-api-details-500"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
...
      "ApisListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDetailsResponse"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "limit",
              "hasMore"
            ],
            "properties": {
              "limit": {
                "type": "integer"
              },
              "hasMore": {
                "type": "boolean"
              },
              "nextCursor": {
                "type": "string",
                "description": "Opaque base64 cursor for the next page. Absent when hasMore is false."
              }
            }
          }
        }
      },
      "ApiCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "base_url",
          "category",
          "endpoints"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "method",
                "price_per_call_usdc"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "method": {
                  "type": "string"
                },
                "price_per_call_usdc": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ApiCreateResponse": {
        "type": "object",
        "required": [
          "id",
          "developer_id",
          "name",
          "base_url",
          "category",
          "status",
          "endpoints"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "developer_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "base_url": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "api_id",
                "path",
                "method",
                "price_per_call_usdc"
              ],
              "properties": {
                "id": {
                  "type": "integer"
                },
                "api_id": {
                  "type": "integer"
                },
                "path": {
                  "type": "string"
                },
                "method": {
                  "type": "string"
                },
                "price_per_call_usdc": {
                  "type": "string"
                },
                "description": {
                  "type": "string",
                  "nullable": true
                },
                "created_at": {
                  "type": "string"
                },
                "updated_at": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ApiDetailsResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "base_url",
          "category",
          "status",
          "developer",
          "endpoints"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "base_url": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "developer": {
            "type": "object",
            "required": [
              "id",
              "name"
            ],
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "api_id",
                "path",
                "method",
                "price_per_call_usdc"
              ],
              "properties": {
                "id": {
                  "type": "integer"
                },
                "api_id": {
                  "type": "integer"
                },
                "path": {
                  "type": "string"
                },
                "method": {
                  "type": "string"
                },
                "price_per_call_usdc": {
                  "type": "string"
                },
                "description": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PAYMENT_REQUIRED",
          "TOO_MANY_REQUESTS",
          "CONFLICT",
          "INTERNAL_SERVER_ERROR",
          "BAD_GATEWAY",
          "SERVICE_UNAVAILABLE",
          "GATEWAY_TIMEOUT",
          "VALIDATION_ERROR",
          "INVALID_BODY",
          "INVALID_QUERY",
          "INVALID_PARAMS",
          "INVALID_VALUE",
          "GATEWAY_AUTH_CONTEXT_MISSING",
          "UPSTREAM_TARGET_BLOCKED",
          "INSUFFICIENT_BALANCE",
          "SOROBAN_RPC_TIMEOUT",
          "SOROBAN_RPC_ERROR",
          "BILLING_DEDUCTION_FAILED",
          "BILLING_REQUEST_NOT_FOUND",
          "DEVELOPER_NOT_FOUND",
          "API_ACCESS_FORBIDDEN",
          "API_KEY_NOT_FOUND",
          "API_KEY_FORBIDDEN",
          "MISSING_REFRESH_TOKEN",
          "INVALID_REFRESH_TOKEN",
          "REVOKED_TOKEN",
          "EXPIRED_TOKEN",
          "REFRESH_FAILED",
          "REVOKE_FAILED",
          "NOT_AUTHENTICATED",
          "TOKEN_INFO_FAILED",
          "VAULT_NOT_FOUND",
          "VAULT_BALANCE_RETRIEVAL_FAILED",
          "MISSING_AMOUNT",
          "INVALID_AMOUNT_TYPE",
          "INVALID_AMOUNT_FORMAT",
          "INVALID_NETWORK",
          "NETWORK_MISMATCH",
          "INVALID_SOURCE_ACCOUNT",
          "INVALID_TRANSACTION_INPUT",
          "SOURCE_ACCOUNT_NOT_FOUND",
          "INVALID_CONTRACT_ID",
          "NETWORK_UNAVAILABLE",
          "TRANSACTION_BUILD_FAILED",
          "INTERNAL_ERROR",
          "INVALID_WEBHOOK_REGISTRATION",
          "INVALID_WEBHOOK_EVENT_TYPES",
          "WEBHOOK_NOT_FOUND",
          "INVALID_WEBHOOK_URL",
          "WEBHOOK_URL_VALIDATION_FAILED",
          "MISSING_WEBHOOK_SIGNATURE_HEADERS",
          "INVALID_WEBHOOK_TIMESTAMP",
          "WEBHOOK_TIMESTAMP_OUT_OF_WINDOW",
          "MALFORMED_WEBHOOK_SIGNATURE",
          "INVALID_WEBHOOK_SIGNATURE",
          "INVALID_DELIVERY_ID",
          "INVALID_RETRY_POLICY",
          "DLQ_ENTRY_NOT_FOUND",
          "INVALID_IP_FORMAT",
          "IP_NOT_ALLOWED",
          "DATABASE_NOT_AVAILABLE",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "SIMULATION_FAILED",
          "INVALID_AUTH_HEADER",
          "MISSING_TOKEN",
          "INVALID_TOKEN",
          "MISSING_CLAIMS",
          "TOKEN_EXPIRED",
          "TOKEN_NOT_ACTIVE",
          "QUOTA_REQUEST_NOT_FOUND",
          "QUOTA_REQUEST_ALREADY_RESOLVED",
          "INVALID_QUOTA_REQUEST",
          "REQUEST_TIMEOUT",
          "REQUEST_BODY_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "UNPROCESSABLE_ENTITY",
          "USAGE_AGGREGATE_NOT_FOUND",
          "INVALID_EXPORT_SCHEDULE",
          "EXPORT_SCHEDULE_NOT_FOUND",
          "MISSING_AUTH_FIELDS",
          "AUTH_NOT_IMPLEMENTED",
          "COMPONENT_NOT_CONFIGURED"
        ],
        "description": "Canonical Callora backend error code."
      },
...
    }
  }
}